Move from outbound.repository to gateways
