+++ src/main/java/br/com/fiap/soat/techChallenge/gateways/jpa/ClienteJpaRepository.java
@@ -1,6 +1,6 @@
-package br.com.fiap.soat.techChallenge.application.adapter.outbound.repository.jpa;
+package br.com.fiap.soat.techChallenge.gateways.jpa;
 
-import br.com.fiap.soat.techChallenge.application.adapter.outbound.repository.jpa.entity.ClienteEntity;
+import br.com.fiap.soat.techChallenge.gateways.jpa.entity.ClienteEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
