+++ src/main/java/br/com/fiap/soat/techChallenge/gateways/jpa/PedidoJpaRepository.java
@@ -1,0 +1,9 @@
+package br.com.fiap.soat.techChallenge.gateways.jpa;
+
+import br.com.fiap.soat.techChallenge.gateways.jpa.entity.PedidoEntity;
+import org.springframework.data.repository.CrudRepository;
+
+import java.util.UUID;
+
+public interface PedidoJpaRepository extends CrudRepository<PedidoEntity, UUID> {
+}
