+++ src/main/java/br/com/fiap/soat/techChallenge/gateways/jpa/ProdutoJpaRepository.java
@@ -1,5 +1,5 @@
-package br.com.fiap.soat.techChallenge.application.adapter.outbound.repository.jpa;
-import br.com.fiap.soat.techChallenge.application.adapter.outbound.repository.jpa.entity.ProdutoEntity;
+package br.com.fiap.soat.techChallenge.gateways.jpa;
+import br.com.fiap.soat.techChallenge.gateways.jpa.entity.ProdutoEntity;
 import br.com.fiap.soat.techChallenge.entities.Produto;
 import org.springframework.data.jpa.repository.JpaRepository;
