+++ src/main/java/br/com/fiap/soat/techChallenge/gateways/jpa/entity/ProdutoEntity.java
@@ -1,4 +1,4 @@
-package br.com.fiap.soat.techChallenge.application.adapter.outbound.repository.jpa.entity;
+package br.com.fiap.soat.techChallenge.gateways.jpa.entity;
 import br.com.fiap.soat.techChallenge.entities.Produto;
 import jakarta.persistence.*;
