+++ src/main/java/br/com/fiap/soat/techChallenge/gateways/jpa/entity/mappers/ProdutoMapper.java
@@ -1,6 +1,6 @@
-package br.com.fiap.soat.techChallenge.application.adapter.outbound.repository.jpa.entity.mappers;
+package br.com.fiap.soat.techChallenge.gateways.jpa.entity.mappers;
 
-import br.com.fiap.soat.techChallenge.application.adapter.outbound.repository.jpa.entity.ProdutoEntity;
+import br.com.fiap.soat.techChallenge.gateways.jpa.entity.ProdutoEntity;
 import br.com.fiap.soat.techChallenge.entities.Produto;
 
 public class ProdutoMapper {
